book-image-classifier/backend/requirements.txt
Flask
Flask-CORS
Pillow
opencv-python
pytesseract
pandas
numpy
scikit-image
python-multipart
Werkzeug